cli: implement help <command>
Allow for calling `snapcraft help <comand>.

Closes: #1699
