+++ snapcraft/cli/__main__.py
@@ -14,6 +14,6 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from . import run
+from ._runner import run
 
 run(prog_name='snapcraft')
